Update docs/languages/en/user-guide/styling-and-translations.rst
Corrections:
"Press Save in the toolbar and poedit will create an ``en_US.mo`` file for us."

Poedit doesn't create an 'en.US.mo' file. The file already exists in the Skeleton Application. 

It'd be nice to know if it modifies the file instead. Lastly the translation isn't working on Windows.

Thanks