Updated index for ZF2 docs.
The user-guide section has been folded into a single section of
'Learning Zend Framework 2' so that all tutorials are together.

The Learning DI tutorial has been moved into a subsection of
'Learning Zend Framework 2' called Concepts and best practices
as it is not really a tutorial, and this name is a better fit,
and allows us to add best practice docs, which people have
been requesting in the docs comments.