Update config.advanced.rst
Refactored the text a little bit to reflect the change with my configuration mapping table. The title to me could use a better wording maybe. Additionally the "Configuration Priority" should be renamed and ideally rewritten a little bit as this basically  belongs to the Configuration merging workflow imo... input required