Update to Installation using Composer
I have updated the docs in the 'Installation using Composer' section as I had to do a couple more steps that are not listed in the docs. I have also updated the example command shown. Is this OK? Let me know if it is not and I will make amends. Thanks.