Update to Installing Zend Tool using Composer
I have updated the docs for installing Zend Tool using Composer. I had to do a couple more steps that are not listed in the docs. I have also updated the example command shown. Is this OK? Let me know if it is not and I will make amends. Thanks.
NOTE: please delete my last pull request, it is wrong -sorry.