Added Composer Package Name
I had a difficult time locating the correct value to use in Composer to require this package. It might be common knowledge with experienced ZF devs, so, my proposed change might not be helpful to many. For me, though, as a newbie to ZF, having the composer package name by the discussion of installing it with Composer is helpful. Thanks!